test (PurchaseAmountValidatorTest): 구입 금액 검증 테스트 이동
